refactor: Update className and button.configs
Separate className from object props in configs for enhanced efficiency.
Rename button.configs to signInButton.configs for clarity.
