+++ app/_components/ui/button/signInButton/index.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { styleButton } from '../button.styles';
 import { PropsSignInButton } from '../button.types';
 import { ButtonWithTooltip } from '../buttonWithTooltip';
 import { signIn } from 'next-auth/react';
@@ -10,6 +11,7 @@ export const SignInButton = ({ configs = {} }: PropsSignInButton) => {
   return (
     <ButtonWithTooltip
       configs={configs}
+      className={styleButton({ className: 'max-ml:mb-3' })}
       onClick={() => signIn()}
     >
       {buttonName}
