+++ app/_components/ui/button/signInButton/signInButton.configs.ts
@@ -1,14 +1,10 @@
 import { createConfigs } from '@/_lib/utils/configs.utils';
-import { styleButton } from './button.styles';
 
 export const configsSignInButton = createConfigs({
   options: {
     buttonName: {
       signIn: 'Sign in',
       getStarted: 'Get started',
-    },
-    className: {
-      signIn: styleButton({ className: 'max-ml:mb-3' }),
     },
     tooltip: {
       signIn: 'Sign in',
@@ -22,13 +18,11 @@ export const configsSignInButton = createConfigs({
   presetOptions: {
     getStarted: {
       buttonName: 'getStarted',
-      className: 'signIn',
       tooltip: 'getStarted',
     },
   },
   defaultOptions: {
     buttonName: 'signIn',
-    className: 'signIn',
     tooltip: 'signIn',
   },
 });
